requirements.txt
pywebostv==0.8.4
google-cloud-speech==1.3.1
sounddevice==0.3.14
scipy==1.4.1
pydub==0.23.1

